backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.9.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.9.0
python-multipart==0.0.6
